third_party/blink/web_tests/http/tests/security/naked-global-object-expected.txt

PASS undefined is undefined.
PASS successfullyParsed is true

TEST COMPLETE
